+++ setup.cfg
@@ -1,0 +1,55 @@
+# This file is used to configure your project.
+# Read more about the various options under:
+# http://setuptools.readthedocs.io/en/latest/setuptools.html#configuring-setup-using-setup-cfg-files
+
+[metadata]
+name = touchstone
+description = To store benchmark results in elasticsearch for long term investigations.
+author =
+author-email =
+license = mit
+long-description = file: README.md
+long-description-content-type = text/markdown; charset=UTF-8
+classifiers =
+    Development Status :: 4 - Beta
+    Programming Language :: Python
+
+[options]
+zip_safe = False
+packages = find:
+include_package_data = True
+package_dir =
+    =src
+# Add here dependencies of your project (semicolon/line-separated), e.g.
+install_requires = configparser; elasticsearch>=6.0.0,<=7.0.2; statistics; numpy; pyyaml; requests; redis
+# tests_require = pytest; pytest-cov
+# Require a specific Python version, e.g. Python 2.7 or >= 3.4
+# python_requires = >=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*
+
+[options.packages.find]
+where = src
+
+[options.extras_require]
+# Add here additional requirements for extra features, to install with:
+# `pip install touchstone[PDF]` like:
+# PDF = ReportLab; RXP
+
+[options.entry_points]
+# Add here console scripts like:
+console_scripts =
+    run_snafu = snafu.run_snafu:render
+
+[aliases]
+dists = bdist_wheel
+
+[bdist_wheel]
+# Use this option if your package is pure-python
+universal = 1
+
+[flake8]
+# Some sane defaults for the code style checker flake8
+exclude =
+    .tox
+    build
+    dist
+    .eggs
