+++ src/snafu/ci/common.sh
