+++ src/snafu/hammerdb/Dockerfile
