+++ src/snafu/hammerdb/ci_test.sh
