+++ src/snafu/hammerdb/uid_entrypoint
