+++ src/snafu/image_resources/centos8-appstream.repo
