+++ src/snafu/image_resources/centos8.repo
