+++ src/snafu/linpack_wrapper/Dockerfile
