+++ src/snafu/linpack_wrapper/linpack.sh
