+++ src/snafu/smallfile_wrapper/Dockerfile
