+++ src/snafu/smallfile_wrapper/ci_test.sh
