+++ src/snafu/smallfile_wrapper/sync-diagram.png
