+++ src/snafu/uperf_wrapper/uperf_wrapper.py
@@ -14,7 +14,7 @@
 import os
 import argparse
 
-from uperf_wrapper.trigger_uperf import Trigger_uperf
+from .trigger_uperf import Trigger_uperf
 
 
 class uperf_wrapper():
