+++ src/snafu/utils/wrapper_factory.py
@@ -1,14 +1,14 @@
 
 # from backpack_wrapper.backpack_wrapperimport backpack_wrapper
-from fio_wrapper.fio_wrapper import fio_wrapper
-from smallfile_wrapper.smallfile_wrapper import smallfile_wrapper
-from uperf_wrapper.uperf_wrapper import uperf_wrapper
-from pgbench_wrapper.pgbench_wrapper import pgbench_wrapper
-from fs_drift_wrapper.fs_drift_wrapper import fs_drift_wrapper
-from cluster_loader.cluster_loader import cluster_loader_wrapper
-from hammerdb.hammerdb_wrapper import hammerdb_wrapper
-from ycsb_wrapper.ycsb_wrapper import ycsb_wrapper
-from vegeta_wrapper.vegeta_wrapper import vegeta_wrapper
+from ..vegeta_wrapper.vegeta_wrapper import vegeta_wrapper
+from ..fio_wrapper.fio_wrapper import fio_wrapper
+from ..smallfile_wrapper.smallfile_wrapper import smallfile_wrapper
+from ..uperf_wrapper.uperf_wrapper import uperf_wrapper
+from ..pgbench_wrapper.pgbench_wrapper import pgbench_wrapper
+from ..fs_drift_wrapper.fs_drift_wrapper import fs_drift_wrapper
+from ..cluster_loader.cluster_loader import cluster_loader_wrapper
+from ..hammerdb.hammerdb_wrapper import hammerdb_wrapper
+from ..ycsb_wrapper.ycsb_wrapper import ycsb_wrapper
 
 import logging
 logger = logging.getLogger("snafu")
