+++ src/snafu/vfs_stat.py
